backend/requirements.txt
# Django Backend Requirements for MobileFix Pro
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File Handling
Pillow==10.1.0
python-magic==0.4.27
PyPDF2==3.0.1

# Payment Gateways
razorpay==1.4.1
stripe==7.8.0

# API & Validation
django-filter==23.5
django-rest-swagger==2.2.0
drf-yasg==1.21.7

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Utils
requests==2.31.0
celery==5.3.4
redis==5.0.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
